+++ samples/s11-imgui/hostCode.cpp
@@ -181,7 +181,7 @@ main(int ac, char **av) {
   guiData->vertex = gprtBufferGetHandle<float3>(guiVertexBuffer);
   guiData->index = gprtBufferGetHandle<int3>(guiIndexBuffer);
   guiData->texture = gprtTextureGetHandle(guiColorAttachment);
-  guiData->resolution = float2(fbSize.x, fbSize.y);
+  guiData->resolution = float2((float)fbSize.x, (float)fbSize.y);
 
   gprtBuildShaderBindingTable(context, GPRT_SBT_RASTER);
 
@@ -221,8 +221,8 @@ main(int ac, char **av) {
       // step 1 : Calculate the amount of rotation given the mouse movement.
       float deltaAngleX = (2 * M_PI / fbSize.x);
       float deltaAngleY = (M_PI / fbSize.y);
-      float xAngle = (lastxpos - xpos) * deltaAngleX;
-      float yAngle = (lastypos - ypos) * deltaAngleY;
+      float xAngle = float(lastxpos - xpos) * deltaAngleX;
+      float yAngle = float(lastypos - ypos) * deltaAngleY;
 
       // step 2: Rotate the camera around the pivot point on the first axis.
       float4x4 rotationMatrixX = rotation_matrix(rotation_quat(lookUp, xAngle));
